Changed the tool's structure to generalize
